Separate bucket implementations, clarify README
